+++ src/components/MiradorViewer.js
@@ -1,5 +1,6 @@
 import React, { useEffect } from "react";
 import Mirador from "mirador";
+import "../styles/mirador-custom.css";
 
 // Import annotation plugin from node_modules
 import annotationPlugins from "mirador-annotations";
@@ -7,7 +8,7 @@ import annotationPlugins from "mirador-annotations";
 // Import the SimpleAnnotationServerV2Adapter
 import SimpleAnnotationServerV2Adapter from "./SimpleAnnotationServerV2Adapter";
 
-const MiradorComponent = () => {
+const MiradorComponent = ({ theme }) => {
   useEffect(() => {
     // Configuration for Mirador viewer
     const config = {
@@ -34,14 +35,17 @@ const MiradorComponent = () => {
           imageToolsEnabled: true,
           // Manifest ID of the IIIF manifest to load
           manifestId:
-            "https://media.nga.gov/public/manifests/nga_highlights.json",
+            // "https://media.nga.gov/public/manifests/nga_highlights.json",
+            // "http://127.0.0.1:8080/manifest_structures.json", // TEST MANIFEST v2 ON "SIMPLE WEB SERVER"
+            "https://raw.githubusercontent.com/sterenz/TEST_MANIFEST/main/manifests/manifest_structures_v3.json", // TEST MANIFEST v3 ON "SIMPLE WEB SERVER"
         },
       ],
+      ...theme, // Merge the theme configuration
     };
 
     // Initialize Mirador viewer with configuration and annotation plugins
     Mirador.viewer(config, [...annotationPlugins]);
-  }, []); // Run effect only once when component mounts
+  }, [theme]); // Run effect when theme changes
 
   // Render the container for Mirador viewer
   return <div id="mirador-container"></div>;